Moves resource constants to different module
