+++ haskanoid.cabal
@@ -130,6 +130,7 @@ executable haskanoid
     other-modules: DisplaySDL2
                    Resources
                    ResourceManagerSDL2
+                   ResourcesSDL2
     extra-libraries: SDL2_image SDL2_ttf SDL2_mixer SDL2
 
   if flag(ghcjs)
